# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Smart City Management System resources",
        "title": "Smart City Management System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/emergency/alerts/areas/{areaId}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send an emergency alert to a specific area.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emergency"
                ],
                "summary": "Send an alert to an area",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Area ID",
                        "name": "areaId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Alert Data",
                        "name": "alert",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emergency_response.AlertsCreateReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/emergency_response.AlertsRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/emergency/evacuation/routes/{areaId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get recommended evacuation routes for a specific area.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emergency"
                ],
                "summary": "Get evacuation routes for an area",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Area ID",
                        "name": "areaId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/emergency_response.RoutesGetByIdRes"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/emergency/incidents": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Report a new incident with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emergency"
                ],
                "summary": "Report a new incident",
                "parameters": [
                    {
                        "description": "Incident Data",
                        "name": "incident",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emergency_response.IncidentsCreateReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/emergency_response.IncidentsRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/emergency/incidents/active": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all currently active incidents.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emergency"
                ],
                "summary": "Get active incidents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/emergency_response.IncidentsGetByIdRes"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/emergency/incidents/{incidentId}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the status of an incident by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emergency"
                ],
                "summary": "Update incident status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Incident ID",
                        "name": "incidentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Incident Status Update",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emergency_response.IncidentsUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emergency_response.IncidentsRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/emergency/resources/available/{emergencyTypeId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all available resources for a specific type of emergency.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emergency"
                ],
                "summary": "Get available resources by emergency type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Emergency Type ID",
                        "name": "emergencyTypeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/emergency_response.ResourcesGetByIdRes"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/emergency/resources/{resourceId}/dispatch": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Dispatch a resource to an incident or location.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emergency"
                ],
                "summary": "Dispatch a resource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource ID",
                        "name": "resourceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Resource Dispatch Data",
                        "name": "dispatch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emergency_response.DispatchesCreateReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/emergency_response.DispatchesRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "emergency_response.AlertsCreateReq": {
            "type": "object",
            "properties": {
                "affected_area": {
                    "type": "string"
                },
                "issued_at": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "emergency_response.AlertsRes": {
            "type": "object",
            "properties": {
                "affected_area": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "issued_at": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "emergency_response.DispatchesCreateReq": {
            "type": "object",
            "properties": {
                "arrived_at": {
                    "type": "string"
                },
                "dispatched_at": {
                    "type": "string"
                },
                "incident_id": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "string"
                }
            }
        },
        "emergency_response.DispatchesRes": {
            "type": "object",
            "properties": {
                "arrived_at": {
                    "type": "string"
                },
                "dispatched_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "incident_id": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "string"
                }
            }
        },
        "emergency_response.IncidentsCreateReq": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "reported_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "emergency_response.IncidentsGetByIdRes": {
            "type": "object",
            "properties": {
                "incident": {
                    "$ref": "#/definitions/emergency_response.IncidentsRes"
                }
            }
        },
        "emergency_response.IncidentsRes": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "reported_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "emergency_response.IncidentsUpdateReq": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "incident": {
                    "$ref": "#/definitions/emergency_response.IncidentsCreateReq"
                }
            }
        },
...
        "emergency_response.ResourcesGetByIdRes": {
            "type": "object",
            "properties": {
                "resource": {
                    "$ref": "#/definitions/emergency_response.ResourcesRes"
                }
            }
        },
        "emergency_response.ResourcesRes": {
            "type": "object",
            "properties": {
                "current_location": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "emergency_response.RoutesGetByIdRes": {
            "type": "object",
            "properties": {
                "route": {
                    "$ref": "#/definitions/emergency_response.RoutesRes"
                }
            }
        },
        "emergency_response.RoutesRes": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "end_point": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "start_point": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}